Added background color change for day setting compatibility
