+++ OpenHouse/OpenHouse.sln
@@ -1,0 +1,17 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenHouse", "OpenHouse.csproj", "{C869B363-BDBA-4520-9264-E38A796658C4}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C869B363-BDBA-4520-9264-E38A796658C4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C869B363-BDBA-4520-9264-E38A796658C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C869B363-BDBA-4520-9264-E38A796658C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C869B363-BDBA-4520-9264-E38A796658C4}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
