--- Room.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace OpenHouse
-{
-    class Room
-    {
-        public string Name { get; set; }
-        public bool HasBeenChecked { get; set; }
-        public string Description { get; set; }
-
-        public Room()
-        {
-
-        }
-    }
-}
